requirements/testing.txt
pytest>=2.7.0
pytest-django>=2.8.0
pytest-describe>=0.10.0
pytest-mock>=0.4.1
